+++ app/components/charts/comparison_chart_component/comparison_chart_component.html.erb
