+++ app/views/comparisons/base/_charts.html.erb
@@ -1,3 +1,3 @@
 <% @charts.each do |chart| %>
-  <%= render ComparisonChartComponent.new(**chart) %>
+  <%= render Charts::ComparisonChartComponent.new(**chart) %>
 <% end %>
